feat: :sparkles: ensures that the user is logged in on all routes

+++ src/routes/memories.ts
@@ -1,11 +1,17 @@
 import { FastifyInstance } from 'fastify'
 import { z } from 'zod'
-
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  // ensures that the user is logged in
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  // list memories
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: { userId: request.user.sub },
       orderBy: { createAt: 'asc' },
     })
     return memories.map((memory) => {
@@ -17,7 +23,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  // list memories by id
+  app.get('/memories/:id', async (request, replay) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -28,9 +35,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
       where: { id },
     })
 
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return replay.status(401).send()
+    }
+
     return memory
   })
 
+  // add memorie
   app.post('/memories/:id', async (request) => {
     const bodySchema = z.object({
       content: z.string(),
@@ -45,17 +57,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '6a2f567f-d9d6-4a69-9263-0331c07644d1',
+        userId: request.user.sub,
       },
     })
 
     return memory
   })
-
-  app.put('/memories/:id', async (request) => {
+  // update memorie
+  app.put('/memories/:id', async (request, replay) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
+
     const { id } = paramsSchema.parse(request.params)
 
     const bodySchema = z.object({
@@ -66,7 +79,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return replay.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: { content, coverUrl, isPublic },
     })
@@ -74,11 +95,21 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  // delete memorie
+  app.delete('/memories/:id', async (request, replay) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return replay.status(401).send()
+    }
+
     await prisma.memory.delete({ where: { id } })
   })
 }
